Summary: Changes in the API 'routes.py' script to implement security measures related with user roles and schema accessibility, new user creation, user promotion and user account deletion. Simplification and speed improvements to endpoints and functions used to add or link alleles. Code modifications to improve readability, reduce redundancy and improve control flow.
Main changes:
- Most of the SPARQL queries used throughout the 'routes.py' script
have been moved to the 'sparql_queries.py' in order to reduce the
number of times certain queries were repeated. Collecting all SPARQL
queries in a single file also improves code readability and helps
managing any existing or new SPARQL queries.

- Wrote a 'generate' function to substitute similar functions that
were being defined throughout the code.

- The process of adding or liking alleles has been simplified by
removing verifications and guaranteeing that the necessary verifications
were performed sooner. The process will also retry up to three times
when an allele could not be inserted the first time.

- Changed 'create_role' function that runs before the first request
to the application so that it does not delete all users when the
application reloads the script.

- Added checks to several endpoints to verify that user exists both
in Postgres and Virtuoso databases.

- New users are added to Postgres and Virtuoso and promoting an user
also applies changes to both databases.

- Endpoint used by the Admin to create new users was hidden and can
create users with the role of User or Contributor.

- Endpoint to delete user checks role and does not allow to delete
Admin. Will check if user to delete exists. Deletes user data from
Postgres and Virtuoso.

- New schema properties: clustering word_size, cluster similarity
threshold, representative similarity exclusion threshold, intra
cluster similarity exclusion threshold, schema locking property.

- Created schema locking endpoints to get locking status of schema
and change locking status.

- Modified the code in endpoints that accept option to filter results
by species, DNA sequence or other input parameters so that the
queries are correctly modified and the correct results are displayed.

- Code modifications to improve readability and reduce the number
of lines through code simplification or removal of redundant code.

- Chose to adopt if/else control flow where try/except clauses are
not necessary or to reduce the number of exceptions that might be
raised and not properly caught or explained.

- Increased number of security checks where necessary (related with
user role or schema locking) and removed user role checks in endpoints
where it can be controled through the custom decorators that were
previously defined.